# schema/openapi.yaml
openapi: 3.1.0
info:
  title: Calendar API
  version: 0.0.0
tags:
  - name: EventTypes
  - name: Slots
  - name: Bookings
  - name: Owner
  - name: AdminEventTypes
paths:
  /api/admin/event-types:
    post:
      operationId: AdminEventTypes_create
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar.Domain.EventType'
      tags:
        - AdminEventTypes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventTypeCreateInput'
  /api/admin/event-types/{id}:
    patch:
      operationId: AdminEventTypes_update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar.Domain.EventType'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
      tags:
        - AdminEventTypes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventTypeUpdateInput'
  /api/bookings:
    post:
      operationId: Bookings_create
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar.Domain.Booking'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictResponse'
      tags:
        - Bookings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreateInput'
    get:
      operationId: Bookings_list
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Calendar.Domain.Booking'
      tags:
        - Bookings
  /api/event-types:
    get:
      operationId: EventTypes_list
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Calendar.Domain.EventType'
      tags:
        - EventTypes
  /api/event-types/{eventTypeId}/slots:
    get:
      operationId: Slots_list
      parameters:
        - name: eventTypeId
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: false
          schema:
            type: string
          explode: false
        - name: to
          in: query
          required: false
          schema:
            type: string
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Calendar.Domain.Slot'
      tags:
        - Slots
  /api/event-types/{id}:
    get:
      operationId: EventTypes_read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar.Domain.EventType'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
      tags:
        - EventTypes
  /api/owner:
    get:
      operationId: OwnerOps_get
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar.Domain.Owner'
      tags:
        - Owner
components:
  schemas:
    BookingCreateInput:
      type: object
      required:
        - eventTypeId
        - slotStart
        - slotEnd
      properties:
        eventTypeId:
          type: string
        slotStart:
          type: string
        slotEnd:
          type: string
        guestName:
          type: string
        guestContact:
          type: string
    BookingCreateInputExample:
      type: object
      required:
        - eventTypeId
        - slotStart
        - slotEnd
      properties:
        eventTypeId:
          type: string
        slotStart:
          type: string
        slotEnd:
          type: string
        guestName:
          type: string
        guestContact:
          type: string
      examples:
        - eventTypeId: 550e8400-e29b-41d4-a716-446655440000
          slotStart: '2025-06-01T10:00:00Z'
          slotEnd: '2025-06-01T10:30:00Z'
          guestName: Иван Петров
          guestContact: ivan@example.com
    Calendar.Domain.Booking:
      type: object
      required:
        - id
        - eventTypeId
        - slotStart
        - slotEnd
        - createdAt
      properties:
        id:
          type: string
        eventTypeId:
          type: string
        slotStart:
          type: string
        slotEnd:
          type: string
        guestName:
          type: string
        guestContact:
          type: string
        createdAt:
          type: string
      examples:
        - id: 770e8400-e29b-41d4-a716-446655440002
          eventTypeId: 550e8400-e29b-41d4-a716-446655440000
          slotStart: '2025-06-01T10:00:00Z'
          slotEnd: '2025-06-01T10:30:00Z'
          guestName: Иван Петров
          guestContact: ivan@example.com
          createdAt: '2025-05-20T12:00:00Z'
    Calendar.Domain.EventType:
      type: object
      required:
        - id
        - name
        - durationMinutes
        - ownerId
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
          minimum: 1
          maximum: 120
        ownerId:
          type: string
      examples:
        - id: 550e8400-e29b-41d4-a716-446655440000
          name: Консультация
          description: Индивидуальная консультация
          durationMinutes: 30
          ownerId: 11111111-1111-1111-1111-111111111111
    Calendar.Domain.Owner:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
      examples:
        - id: 11111111-1111-1111-1111-111111111111
          name: Алексей Смирнов
          email: alexey@example.com
    Calendar.Domain.Slot:
      type: object
      required:
        - id
        - eventTypeId
        - startAt
        - endAt
        - status
      properties:
        id:
          type: string
        eventTypeId:
          type: string
        startAt:
          type: string
        endAt:
          type: string
        status:
          type: string
          enum:
            - free
            - booked
      examples:
        - id: 660e8400-e29b-41d4-a716-446655440001
          eventTypeId: 550e8400-e29b-41d4-a716-446655440000
          startAt: '2025-06-01T10:00:00Z'
          endAt: '2025-06-01T10:30:00Z'
          status: free
    ConflictResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    ErrorResponseExample:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
      examples:
        - code: 404
          message: Resource not found
    EventTypeCreateInput:
      type: object
      required:
        - name
        - durationMinutes
      properties:
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
          minimum: 1
          maximum: 120
    EventTypeCreateInputExample:
      type: object
      required:
        - name
        - durationMinutes
      properties:
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
          minimum: 1
          maximum: 120
      examples:
        - name: Консультация
          description: Индивидуальная консультация 30 минут
          durationMinutes: 30
    EventTypeUpdateInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
          minimum: 1
          maximum: 120
    EventTypeUpdateInputExample:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
          minimum: 1
          maximum: 120
      examples:
        - name: Консультация (обновлённая)
          durationMinutes: 45
    NotFoundResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
servers:
  - url: http://localhost:3000
    description: Development server
    variables: {}